Enable checkpoint load from HF
